requirements.txt
Llama-index
pypdf
streamlit~=1.39.0
langchain_groq
python-dotenv~=1.0.1
langchain~=0.3.3
langserve
fastapi
uvicorn